Optimization: don't need to loop words in group
By construction, groups store only words that are anagrams. Thus,
to check if a word belongs to a group, we don't need to check if
it's an anagram of all words in the group, not even of any word
in the group, but just the first word in the group!

We'd gained less than 10% in performance, but this was an obvious
fix for a stupid algorithm.