Fix false positive in isAnagram
Since we were just checking length of strings and that all chars
in one belongs to the other, we had some false positives, like
'stress' and 'treses', which aren't anagrams but pass those checks.

We fixed it by ordering the chars in a string to get some kind of
canonical expresion, and then comparing those expresions. For
instance, in the previous example, we get 'erssst' and 'eersst',
which aren't equal.