Optimization attempt on sorting chars
Sorting chars within a string in PHP has the little problem that
a string is not an array, and we end up setting up 2 or 3 loops
to transform into array, sort and transform back again into
string. Of course this is inneficient and we'll probably must
consider modeling a word as an array of chars and reading that
directly from the input.

But in the meanwhile, we've made an attempt to improve sorting
based in the fact that chars in a word belong to a very little
set of values. Thus, we applied the 'cubes' sorting method instead
of php arrays "sort" one. The 'cubes' algorithm is O(N) since it
takes the time of "reading" the N elements, but since the words
aren't very long we don't see an improvement in performance.